dev (#3629)
* refactor(gapi): rewrote user journey code

* refactor(gapi): refactored logs

* refactor(gapi): transform userJourney data to sunburst format

* refactor(gapi): applied requested changes to userJourney

* refactor(gapi): changed main analytics for local dev

* refactor(gapi): use $current_path instead of $properties.url_path
refactor(DB): defined $current_path

* refactor(gapi): validate product analytics payload
refactor(gapi): support exclude for userJourney
refactor(gapi): cleaned code for userJourney

* refactor(gapi): userJourney decode which table to use
refactor(gapi): userJourney changed sunburst response structure

* refactor(gapi): timeseries return empty response

* refactor(gapi): funnels join sessions table only when required

* refactor(gapi): heatmaps use $current_path

+++ backend/pkg/analytics/charts/metric_table_errors.go
@@ -73,7 +73,7 @@ func (t *TableErrorsQueryBuilder) buildQuery(p *Payload) (string, error) {
 		density = 7
 	}
 	durMs := p.EndTimestamp - p.StartTimestamp
-	stepMs := durMs / int64(density-1)
+	stepMs := int64(durMs) / int64(density-1)
 	startMs := (p.StartTimestamp / 1000) * 1000
 	endMs := (p.EndTimestamp / 1000) * 1000
